Added queue interface
Modified slist slightly by changing size to _size to indicate it should not be modified by user.
Modified slist_free() to take a custom function for freeing the data if that option is chosen. This is necessary in order to handle pointers to structs that may have dynamically allocated memory variables.